+++ server/lib/rdio-scanner/config.js
@@ -157,35 +157,61 @@ export class Config extends EventEmitter {
 
     get options() {
         return {
-            autoPopulate: this._options.autoPopulate ?? defaults.options.autoPopulate,
-            dimmerDelay: this._options.dimmerDelay ?? defaults.options.dimmerDelay,
-            disableAudioConversion: this._options.disableAudioConversion ?? defaults.options.disableAudioConversion,
-            disableDuplicateDetection: this._options.disableDuplicateDetection ?? defaults.options.disableDuplicateDetection,
-            duplicateDetectionTimeFrame: this._options.duplicateDetectionTimeFrame ?? defaults.options.duplicateDetectionTimeFrame,
-            keypadBeeps: this._options.keypadBeeps ?? defaults.options.keypadBeeps,
-            pruneDays: this._options.pruneDays ?? defaults.options.pruneDays,
-            sortTalkgroups: this._options.sortTalkgroups ?? defaults.options.sortTalkgroups,
+                autoPopulate: typeof this._options.autoPopulate === 'boolean'
+                    ? this._options.autoPopulate
+                    : defaults.options.autoPopulate,
+                dimmerDelay: typeof this._options.dimmerDelay === 'number'
+                    ? this._options.dimmerDelay
+                    : defaults.options.dimmerDelay,
+                disableAudioConversion: typeof this._options.disableAudioConversion === 'boolean'
+                    ? this._options.disableAudioConversion
+                    : defaults.options.disableAudioConversion,
+                disableDuplicateDetection: typeof this._options.disableDuplicateDetection === 'boolean'
+                    ? this._options.disableDuplicateDetection
+                    : defaults.options.disableDuplicateDetection,
+                duplicateDetectionTimeFrame: typeof this._options.duplicateDetectionTimeFrame === 'number'
+                    ? this._options.duplicateDetectionTimeFrame
+                    : defaults.options.duplicateDetectionTimeFrame,
+                keypadBeeps: typeof this._options.keypadBeeps === 'string'
+                    ? this._options.keypadBeeps
+                    : defaults.options.keypadBeeps,
+                pruneDays: typeof this._options.pruneDays === 'number'
+                    ? this._options.pruneDays
+                    : defaults.options.pruneDays,
+                sortTalkgroups: typeof this._options.sortTalkgroups === 'boolean'
+                    ? this._options.sortTalkgroups
+                    : defaults.options.sortTalkgroups,
         };
     }
 
     set options(options) {
         if (options !== null && !Array.isArray(options) && typeof options === 'object') {
-            this._options = Object.keys(options).reduce((o, k) => {
-                if ([
-                    'autoPopulate',
-                    'dimmerDelay',
-                    'disableAudioConversion',
-                    'disableDuplicateDetection',
-                    'duplicateDetectionTimeFrame',
-                    'keypadBeeps',
-                    'pruneDays',
-                    'sortTalkgroups',
-                ].includes(k)) {
-                    o[k] = options[k] ?? defaults.options[k];
-                }
-
-                return o;
-            }, {});
+            this._options = {
+                autoPopulate: typeof options.autoPopulate === 'boolean'
+                    ? options.autoPopulate
+                    : defaults.options.autoPopulate,
+                dimmerDelay: typeof options.dimmerDelay === 'number'
+                    ? options.dimmerDelay
+                    : defaults.options.dimmerDelay,
+                disableAudioConversion: typeof options.disableAudioConversion === 'boolean'
+                    ? options.disableAudioConversion
+                    : defaults.options.disableAudioConversion,
+                disableDuplicateDetection: typeof options.disableDuplicateDetection === 'boolean'
+                    ? options.disableDuplicateDetection
+                    : defaults.options.disableDuplicateDetection,
+                duplicateDetectionTimeFrame: typeof options.duplicateDetectionTimeFrame === 'number'
+                    ? options.duplicateDetectionTimeFrame
+                    : defaults.options.duplicateDetectionTimeFrame,
+                keypadBeeps: typeof options.keypadBeeps === 'string'
+                    ? options.keypadBeeps
+                    : defaults.options.keypadBeeps,
+                pruneDays: typeof options.pruneDays === 'number'
+                    ? options.pruneDays
+                    : defaults.options.pruneDays,
+                sortTalkgroups: typeof options.sortTalkgroups === 'boolean'
+                    ? options.sortTalkgroups
+                    : defaults.options.sortTalkgroups,
+            };
 
             this.emit('changes');
         }